# updating the build with SGX