--- a/IDE/LINUX-SGX/sgx_t_static.mk
+++ b/IDE/LINUX-SGX/sgx_t_static.mk
@@ -95,13 +95,14 @@ Wolfssl_C_Files :=$(WOLFSSL_ROOT)/wolfcrypt/src/aes.c\
 					$(WOLFSSL_ROOT)/src/ssl.c\
 					$(WOLFSSL_ROOT)/src/tls.c\
 					$(WOLFSSL_ROOT)/wolfcrypt/src/wc_encrypt.c\
-					$(WOLFSSL_ROOT)/wolfcrypt/src/wolfevent.c\
+					$(WOLFSSL_ROOT)/wolfcrypt/src/wolfevent.c
+
 
 ifeq ($(HAVE_WOLFSSL_ASSEMBLY), 1)
-    Wolfssl_ASM_Files := $(WOLFSSL_ROOT)/wolfcrypt/src/aes_asm.asm\
+	Wolfssl_ASM_Files := $(WOLFSSL_ROOT)/wolfcrypt/src/aes_asm.asm\
                $(WOLFSSL_ROOT)/wolfcrypt/src/sp_x86_64_asm.asm
 
-    Wolfssl_S_Files := $(WOLFSSL_ROOT)/wolfcrypt/src/aes_asm.S\
+	Wolfssl_S_Files := $(WOLFSSL_ROOT)/wolfcrypt/src/aes_asm.S\
                $(WOLFSSL_ROOT)/wolfcrypt/src/aes_gcm_asm.S\
                $(WOLFSSL_ROOT)/wolfcrypt/src/poly1305_asm.S\
                $(WOLFSSL_ROOT)/wolfcrypt/src/sha256_asm.S\
@@ -113,14 +114,16 @@ ifeq ($(HAVE_WOLFSSL_ASSEMBLY), 1)
                $(WOLFSSL_ROOT)/wolfcrypt/src/sha512_asm.S
 
 
-    Wolfssl_C_Extra_Flags += -DWOLFSSL_X86_64_BUILD\
+	Wolfssl_C_Extra_Flags += -DWOLFSSL_X86_64_BUILD\
                -DWOLFSSL_AESNI\
                -maes -masm=intel
-ifeq ($(HAVE_WOLFSSL_SP), 1)
-    Wolfssl_C_Extra_Flags += -DWOLFSSL_SP_X86_64_ASM\
-			     -DWOLFSSL_SP_X86_64\
-			     -DWOLFSSL_SP_ASM
-endif
+
+#SP assembly needs investigated for use with PIE
+#ifeq ($(HAVE_WOLFSSL_SP), 1)
+#    Wolfssl_C_Extra_Flags += -DWOLFSSL_SP_X86_64_ASM\
+#			     -DWOLFSSL_SP_X86_64\
+#			     -DWOLFSSL_SP_ASM
+#endif
 endif
 
 Wolfssl_Include_Paths := -I$(WOLFSSL_ROOT)/ \
@@ -179,6 +182,7 @@ all: libwolfssl.sgx.static.lib.a
 libwolfssl.sgx.static.lib.a: $(Wolfssl_C_Objects)
 	ar rcs libwolfssl.sgx.static.lib.a $(Wolfssl_C_Objects)
 	@echo "LINK =>  $@"
+	@echo "Built with AES-NI ? $(HAVE_WOLFSSL_ASSEMBLY)"
 
 clean:
 	@rm -f $(WOLFSSL_ROOT)/wolfcrypt/benchmark/*.o $(WOLFSSL_ROOT)/wolfcrypt/test/*.o static_trusted/wolfssl_t.* libwolfssl.sgx.static.lib.a  $(Wolfssl_C_Objects)