https://www.wolfssl.com/forums/topic1535-aesni-and-intel-sgx.html